TASK: Move the findAllIterator to the doctrine repository
