+++ Neos.Flow/Classes/Persistence/Repository.php
@@ -12,7 +12,6 @@ namespace Neos\Flow\Persistence;
  */
 
 use Doctrine\Common\Persistence\ObjectManager;
-use Doctrine\ORM\Internal\Hydration\IterableResult;
 use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Persistence\Exception\IllegalObjectTypeException;
 
@@ -121,44 +120,6 @@ abstract class Repository implements RepositoryInterface
     public function findAll()
     {
         return $this->createQuery()->execute();
-    }
-
-    /**
-     * Find all objects and return an IterableResult
-     *
-     * @return IterableResult
-     */
-    public function findAllIterator()
-    {
-        /** @var \Doctrine\ORM\QueryBuilder $queryBuilder */
-        $queryBuilder = $this->entityManager->createQueryBuilder();
-        return $queryBuilder
-            ->select('entity')
-            ->from($this->getEntityClassName(), 'entity')
-            ->getQuery()->iterate();
-    }
-
-    /**
-     * Iterate over an IterableResult and return a Generator
-     *
-     * This method is useful for batch processing a huge result set.
-     *
-     * @param IterableResult $iterator
-     * @param callable $callback
-     * @return \Generator
-     */
-    public function iterate(IterableResult $iterator, callable $callback = null)
-    {
-        $iteration = 0;
-        foreach ($iterator as $object) {
-            $object = current($object);
-            yield $object;
-            if ($callback !== null) {
-                call_user_func($callback, $iteration, $object);
-            }
-
-            $iteration++;
-        }
     }
 
     /**
